make installation of s3 concurent
The idea was to create the namespace concurrently instead of
sequentially to improve the overall installation time.

Then while I was at it I also installed the vm at the same time as the
namespaces so the vm is always nearly ready once all the namespaces are
created

Turns out that it all worked pretty good and the installation time has
been greatly reduce. The code is also more solid and handle better the
case where a namespace fail to install on a node.
The code now just try to deploy the namespace on another node without
crashing the full install as before

Added some more logging to follow the progress more easily.
And the code looks also better with smaller function and cleaner flow
logic

fixes #115